requirements.txt
aiohttp
typing-extensions
backports.zstd
Brotli
colorlog
colorama

setuptools
python-dotenv